Life cycle simulation of Aedes mosquitoes

Function to simulate population dynamics of Aedes mosquitoes

dynamAedes(
  species = "aegypti",
  intro.eggs = 0,
  intro.adults = 0,
  intro.juveniles = 0,
  scale = "ws",
  intro.cells = NULL,
  ihwv = 1,
  temps.matrix = NULL,
  startd = 1,
  endd = 10,
  cells.coords = NULL,
  lat = 0,
  long = 0,
  road.dist.matrix = NULL,
  avgpdisp = NA,
  intro.year = 2020,
  iter = 1,
  n.clusters = 1,
  cluster.type = "PSOCK",
  sparse.output = FALSE,
  compressed.output = TRUE,
  suffix = NA,
  cellsize = 250,
  maxadisp = 600,
  dispbins = 10,
  verbose = FALSE,
  seeding = FALSE
)

Arguments

species

character. Select what species to model: "aegypti", "albopictus", "japonicus", "koreicus". Default species = "aegypti".

intro.eggs

positive integer. number of introduced eggs, default intro.eggs = 100.

intro.adults

positive integer. number of introduced adults, default zero.

intro.juveniles

positive integer. number of introduced juveniles, default zero.

scale

character. Define the model spatial scale: punctual/weather station "ws", local "lc", or regional "rg". Active and passive dispersal is enabled only for scale = "lc". Default scale = "ws".

intro.cells

positive integer. One or more cells (id) where to introduce the population at local ("lc") scale. If intro.cells=NULL, then a random cell is used for introduction; If intro.cells is a vector of cell ids then a cell is drawn at random from the vector (with repetition) for introduction in each model iteration.

ihwv

positive integer. Larval-habitat water volume, define the volume (L) of water habitat presents in each spatial unit (parametrised with data retrieved from doi:10.1111/1365-2664.12620 ). Default lhwv = 1.

temps.matrix

matrix. A matrix of daily (average) temperatures (in degrees Celsius degree x 1000) used to fit the life cycle rates. This matrix must be organised with the daily temperature observations as columns and the geographic position of the i-grid cell as rows.

startd

positive integer. Day of start of the simulations, referring to the column number of temps.matrix.

endd

positive integer. Day of end of the simulation, referring to the column number of temps.matrix.

cells.coords

matrix. A matrix reporting the spatial coordinates of the temperature observations.

lat

numeric. Latitude value of the area of interested used to derive the photoperiod (and thus the diapause eggs allocation function).

long

numeric. Longitude value of the area of interested used to derive the photoperiod (and thus the diapause eggs allocation function)

road.dist.matrix

matrix. when scale = "lc", defines the matrix containing the distances (in meters) between grid cells intersecting the road network for the mosquito passive dispersal process.

avgpdisp

optional. when scale = "lc", define the average car trip distance for the mosquito passive dispersal process. The value can be set by the users (positive numeric), or the estimates made by Pasaoglu et al. 2012) for the following European countries: France "fra", Germany "deu", Italy "ita", Poland "pol", Spain "esp" and the United Kingdom "uk". The average passive dispersal distance must be smaller than the maximum distance in road.dist.matrix.

intro.year

numeric. Year of the beginning of iterations (for photoperiod calculation).

iter

positive integer. Define the number of model iterations.

n.clusters

positive integer. Defines the number of parallel processes.

cluster.type

character. Defines the type of cluster, default "PSOCK".

sparse.output

logical. The output matrix is optimised for sparse-matrix algebra (e.g. zeros are indexed).

compressed.output

logical. Default TRUE, if FALSE provide abundance for each model's subcompartiment; if FALSE abundances are summed per compartment.

suffix

character. Model output suffix for output RDS.

cellsize

(positive integer. When scale = "lc", defines the minimal distance for the active dispersal kernel and should match the spatial resolution of temps.matrix to avoid inconsistencies. Default cellsize = 250

maxadisp

positive integer. When scale = "lc", defines the maximum daily dispersal, default maxadisp = 600.

dispbins

positive integer. When scale = "lc", defines the resolution of the dispersal kernel, default dispbins = 10.

verbose

logical. if TRUE then an overview of population dynamics is printed in the console.

seeding

logical, default FALSE, if seeding=TRUE a fixed seed is applied for result reproducibility.

Value

Matrix or a list of matrices containing, for each iteration, the number of individuals in each life stage per day (and for each grid cell of the study area if scale="lc" or "rg"). If the argument compressed.output=FALSE (default TRUE), the model returns the daily number of individuals in each life stage sub-compartment.

See also

Author

Matteo Marcantonio marcantoniomatteo@gmail.com, Daniele Da Re daniele.dare@uclouvain.be

Examples

## Run dynmAedes at local scale for 5 days
# Make a toy temperature time series
if (FALSE) {
w <- matrix(seq(20,25,length.out=5),ncol=5)*1000
# Run the model
  dynamAedes(
  species="koreicus", 
  scale="ws",
  intro.eggs=10, 
  ihwv=2, 
  temps.matrix=w, 
  startd=2, 
  endd=5,
  lat=42,
  long=8,
  n.clusters=1, 
  iter=1,
  compressed.output=TRUE)
}