Proportion of successful introductions

Compute the proportion of "successful" introductions.

Usage

get_rates_spatial(
  coords = NULL,
  temps.matrix = NULL,
  species = "aegypti",
  rate_fun = NULL,
  spatial = FALSE,
  rate = TRUE,
  n.clusters = 1
)

Arguments

coords

matrix (optional). A matrix reporting the spatial coordinates of the temperature observations.

temps.matrix

matrix, vector. A matrix or a vector of daily average temperature (in Celsius degree /1000) used to fit the life cycle rates. This matrix must be organised with the daily temperature observations as columns and the geographic position of the i-grid cell as rows.

species

character. Select what species to model: "aegypti", "albopictus", "japonicus", "koreicus". Default species = "aegypti".

rate_fun

character. Define one of the temperature-dependent functions of the model, e.g. ".e.hatch.rate".

spatial

logical. Get a raster as output. Default FALSE

rate

logical. Returns the daily rate. Default TRUE

n.clusters

postive integer. Define the number of parallel processes.

Value

Returns the estimated value of a given temperature-dependent function. If spatial = TRUE, it will return a raster.

Author

Matteo Marcantonio marcantoniomatteo@gmail.com, Daniele Da Re daniele.dare@uclouvain.be