Number of invaded cells

Compute a summary of the number of invaded cells over model iterations

icci(input_sim = NA, eval_date = 0, breaks = c(0.25, 0.5, 0.75))

Arguments

input_sim

matrix. dynamAedes.m compressed output matrix (compressed=TRUE).

eval_date

numeric. Define the day of evaluation; it refers to the column number of the input temperature matrix.

breaks

numeric vector. Quantile breaks, default interquartile range: c(0.25,0.5,0.75).

Value

icci returns quantiles of the distribution of the invaded cell number for the specified. The output should be interpreted according to model spatial scale (i.e. scale='rg' or scale='lc' give different interpretation).

Author

Matteo Marcantonio marcantoniomatteo@gmail.com, Daniele Da Re daniele.dare@uclouvain.be